[backend] fix cyclical import by splitting raw s3 commands from file-storage
