+++ opencti-platform/opencti-graphql/src/database/exclusionListCache.ts
@@ -4,7 +4,7 @@ import { checkExclusionLists, convertIpAddr } from '../utils/exclusionLists';
 import type { AuthContext } from '../types/user';
 import { fullEntitiesList } from './middleware-loader';
 import { SYSTEM_USER } from '../utils/access';
-import { getFileContent } from './file-storage';
+import { getFileContent } from './raw-file-storage';
 import { logApp, PLATFORM_INSTANCE_ID } from '../config/conf';
 import { redisGetExclusionListCache, redisSetExclusionListCache, redisUpdateExclusionListStatus } from './redis';
 import { FunctionalError } from '../config/errors';
