+++ opencti-platform/opencti-graphql/src/modules/disseminationList/disseminationList-domain.ts
@@ -23,7 +23,8 @@ import conf, { logApp } from '../../config/conf';
 import { FunctionalError, UnsupportedError } from '../../config/errors';
 import { checkEnterpriseEdition } from '../../enterprise-edition/ee';
 import { createInternalObject, deleteInternalObject, editInternalObject } from '../../domain/internalObject';
-import { downloadFile, getFileContent, loadFile } from '../../database/file-storage';
+import { downloadFile, getFileContent } from '../../database/raw-file-storage';
+import { loadFile } from '../../database/file-storage';
 import { getEntityFromCache } from '../../database/cache';
 import { ENTITY_TYPE_SETTINGS } from '../../schema/internalObject';
 import { OCTI_EMAIL_TEMPLATE } from '../../utils/emailTemplates/octiEmailTemplate';
