+++ opencti-platform/opencti-graphql/src/modules/support/support-domain.ts
@@ -14,7 +14,8 @@ import {
   TELEMETRY_LOG_FILE_PREFIX,
   TELEMETRY_LOG_RELATIVE_LOCAL_DIR
 } from '../../config/conf';
-import { downloadFile, loadedFilesListing, streamConverter } from '../../database/file-storage';
+import { downloadFile } from '../../database/raw-file-storage';
+import { loadedFilesListing, streamConverter } from '../../database/file-storage';
 import type { EditInput, QuerySupportPackagesArgs, SupportPackageAddInput, SupportPackageForceZipInput } from '../../generated/graphql';
 import { EditOperation, PackageStatus } from '../../generated/graphql';
 import { updateAttribute } from '../../database/middleware';
