+++ opencti-platform/opencti-graphql/tests/01-unit/domain/workspace-test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, it } from 'vitest';
-import { streamToString } from '../../../src/database/file-storage';
+import type { Readable } from 'stream';
+import { streamToString } from '../../../src/database/raw-file-storage';
 import { checkConfigurationImport } from '../../../src/modules/workspace/workspace-domain';
 import { createUploadFromTestDataFile } from '../../utils/testQueryHelper';
 
@@ -12,7 +13,8 @@ describe('workspace', () => {
   it.each(cases)('should verify import version compatibility, given invalid version (%s for error %s)', async (filePath, expectedErrorMessage) => {
     const upload = await createUploadFromTestDataFile(filePath, 'invalid-version.json', 'application/json', 'utf8');
     const readStream = upload.file?.createReadStream();
-    const fileContent = await streamToString(readStream);
+    expect(readStream).toBeDefined();
+    const fileContent = await streamToString(readStream as Readable);
     const parsedData = JSON.parse(fileContent.toString());
     const check = () => {
       checkConfigurationImport('dashboard', parsedData);
@@ -23,7 +25,8 @@ describe('workspace', () => {
   it('should verify import version compatibility, given invalid type', async () => {
     const upload = await createUploadFromTestDataFile('20233010_octi_dashboard_Custom Dash_invalid_type.json', 'invalid-version.json', 'application/json', 'utf8');
     const readStream = upload.file?.createReadStream();
-    const fileContent = await streamToString(readStream);
+    expect(readStream).toBeDefined();
+    const fileContent = await streamToString(readStream as Readable);
     const parsedData = JSON.parse(fileContent.toString());
 
     expect(() => checkConfigurationImport('dashboard', parsedData)).toThrowError('Invalid type. Please import OpenCTI dashboard-type only');
@@ -32,7 +35,8 @@ describe('workspace', () => {
   it('should verify import version compatibility, given valid import', async () => {
     const upload = await createUploadFromTestDataFile('20233010_octi_dashboard_Custom Dash_valid.json', 'invalid-version.json', 'application/json', 'utf8');
     const readStream = upload.file?.createReadStream();
-    const fileContent = await streamToString(readStream);
+    expect(readStream).toBeDefined();
+    const fileContent = await streamToString(readStream as Readable);
     const parsedData = JSON.parse(fileContent.toString());
 
     expect(() => checkConfigurationImport('dashboard', parsedData),).not.toThrowError();
