+++ opencti-platform/opencti-graphql/tests/02-integration/01-database/index-file-test.js
@@ -4,7 +4,7 @@ import { createReadStream } from 'node:fs';
 import { elLoadById } from '../../../src/database/engine';
 import { elIndexFiles, elSearchFiles } from '../../../src/database/file-search';
 import { ADMIN_USER, testContext } from '../../utils/testQuery';
-import { getFileContent } from '../../../src/database/file-storage';
+import { getFileContent } from '../../../src/database/raw-file-storage';
 import { INDEX_FILES } from '../../../src/database/utils';
 import { resetFileIndexing } from '../../../src/domain/indexedFile';
 import { uploadToStorage } from '../../../src/database/file-storage-helper';
