+++ opencti-platform/opencti-graphql/tests/02-integration/02-resolvers/exclusionListResolver-test.ts
@@ -1,7 +1,8 @@
 import { describe, it, expect, beforeAll } from 'vitest';
 import gql from 'graphql-tag';
 import Upload from 'graphql-upload/Upload.mjs';
-import { downloadFile, streamConverter } from '../../../src/database/file-storage';
+import { downloadFile } from '../../../src/database/raw-file-storage';
+import { streamConverter } from '../../../src/database/file-storage';
 import { fileToReadStream } from '../../../src/database/file-storage-helper';
 import { elLoadById } from '../../../src/database/engine';
 import { ADMIN_USER, testContext } from '../../utils/testQuery';
@@ -115,7 +116,7 @@ describe('Exclusion list resolver', () => {
       });
 
       it('should create a file', async () => {
-        const fileStream = await downloadFile(exclusionListFileResponse.file_id);
+        const fileStream = await downloadFile(exclusionListFileResponse.file_id ?? '');
         expect(fileStream).not.toBeNull();
         const data = await streamConverter(fileStream);
         expect(data).toEqual('127.0.0.1\n10.10.0.0\n2.2.2.2');
@@ -134,7 +135,7 @@ describe('Exclusion list resolver', () => {
         expect(fieldPatch?.data?.exclusionListFieldPatch.exclusion_list_values_count).toBe(4);
         expect(fieldPatch?.data?.exclusionListFieldPatch.exclusion_list_file_size).toBe(37);
         // verify that file was modified
-        const fileStream = await downloadFile(exclusionListFileResponse.file_id);
+        const fileStream = await downloadFile(exclusionListFileResponse.file_id ?? '');
         expect(fileStream).not.toBeNull();
         const data = await streamConverter(fileStream);
         expect(data).toEqual('127.0.0.1\n10.10.0.0\n12.10.0.0\n2.2.2.2');
@@ -187,7 +188,7 @@ describe('Exclusion list resolver', () => {
       });
 
       it('should have deleted the file', async () => {
-        const fileStream = await downloadFile(exclusionListFileResponse?.file_id);
+        const fileStream = await downloadFile(exclusionListFileResponse.file_id ?? '');
         expect(fileStream).toBeNull();
       });
     });
