+++ opencti-platform/opencti-graphql/tests/utils/testSetup.js
@@ -2,7 +2,7 @@ import '../../src/modules/index';
 import cacheManager from '../../src/manager/cacheManager';
 import { initializeRedisClients } from '../../src/database/redis';
 import { searchEngineInit } from '../../src/database/engine';
-import { initializeFileStorageClient } from '../../src/database/file-storage';
+import { initializeFileStorageClient } from '../../src/database/raw-file-storage';
 import { initExclusionListCache } from '../../src/database/exclusionListCache';
 import { initLockFork } from '../../src/lock/master-lock';
